Compact log after a certain number of additions.
This change implements compaction after some `n` number of entries.
Since the only log entry that matters is the last one, we can
occcasionally go back and remove extraneous entries. Note that
this implementation is pretty naive and will probably need to be
modified in the future, especially once some benchmarking is done.